+++ src/infrastructure/handler/auth/generic/mapper/mapper.go
@@ -2,6 +2,8 @@ package mapper
 
 import (
 	"github.com/lumialvarez/go-grpc-auth-service/src/infrastructure/handler/grpc/auth/pb"
+	"reflect"
+	"strconv"
 )
 
 type Mapper struct {
@@ -15,5 +17,16 @@ func (m Mapper) ToListRequest(id int64, userName string) pb.ListRequest {
 	if len(userName) > 0 {
 		request.UserName = &userName
 	}
+	return request
+}
+
+func (m Mapper) ToReadNotificationRequest(userId any, notificationId string) pb.ReadNotificationRequest {
+	notificationIdInt, _ := strconv.ParseInt(notificationId, 10, 64)
+
+	request := pb.ReadNotificationRequest{
+		UserId:         reflect.ValueOf(userId).Int(),
+		NotificationId: notificationIdInt,
+	}
+
 	return request
 }